try to fix owlbot error

+++ owlbot.py
@@ -19,6 +19,7 @@ from pathlib import Path
 import shutil
 import subprocess
 
+import os
 import synthtool as s
 from synthtool.languages import php
 from synthtool import _tracked_paths
@@ -29,7 +30,7 @@ logging.basicConfig(level=logging.DEBUG)
 protos = [
     ("api", "api"),
     ("extendedoperations", "cloud"),
-    ("location", "cloud"), 
+    ("location", "cloud"),
     ("logging", "google"), # for the metadata
     ("logging", "cloud"),
     ("iam", "google"), # for the metadata
@@ -62,9 +63,12 @@ s.move("metadata/Google/Iam/V1", "metadata/Iam/V1")
 s.move("metadata/Google/Logging/Type", "metadata/Logging/Type")
 
 # remove owl-bot-staging dir
-#shutil.rmtree(Path(php.STAGING_DIR))
+if os.path.exists(php.STAGING_DIR):
+    shutil.rmtree(Path(php.STAGING_DIR))
+
 # remove the metadata/Google files that we copied
-shutil.rmtree(Path("metadata/Google"))
+if os.path.exists("metadata/Google"):
+    shutil.rmtree(Path("metadata/Google"))
 
 s.replace(
     "src/**/*.php",
